Improve rankings and descriptions

+++ src/modules/process_data/process_individual_county.py
@@ -99,7 +99,7 @@ def process_datatype(
     if clean_rules.get("moving_avg"):
         col_to_avg = clean_rules["moving_avg"]
         df["moving_avg"] = df[col_to_avg].rolling(window=7).mean()
-        df["moving_avg"] = df["moving_avg"].fillna(0).astype("int64")
+        df["moving_avg"] = df["moving_avg"].fillna(0)
 
     if clean_rules.get("moving_avg") and clean_rules.get("moving_avg_per_capita"):
         county_pop = get_county_pop(county_name)
